% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.ontoarray-function.R
\name{generator.ontoarray}
\alias{generator.ontoarray}
\title{Make an ontoarray generator for visible neural network (VNN) modeling}
\usage{
generator.ontoarray(tidy_set, index, batch_size)
}
\arguments{
\item{tidy_set}{TidySet, an ExpressionSet with three tables.}

\item{index}{An integer vector of index to select which ontoarray will be
used for training or testing.}

\item{batch_size}{An integer of how much samples are generated everytime
this function runs. If all samples are generated,this function will loop over
the samples.}
}
\value{
output sample generator, a function for argument of \code{generator}
in \code{fit_generator}, \code{evaluate_generator}, or
\code{predict_generator} function from Keras R package.
}
\description{
This function creates a function that generate a batch of ontoarray for
training or testing a Keras Convolutional Neural Network (CNN) model using
\code{fit_generator}, \code{evaluate_generator}, or \code{predict_generator}
function from Keras R package.
}
\examples{

## Create input example
input=utils.example()

## Compile input to a TidySet
tidy_set=
  TidySet.compile(
    value=input$value
    ,outcome=input$outcome
    ,similarity=input$similarity
    ,mapping=input$mapping
    ,ontology=input$ontology
  )

## Create ontonet (Keras model object) generator function
ontonet=generator.ontonet(tidy_set)

## Randomize sample and split indices for train and test set
set.seed(33)
index=sample(1:dim(tidy_set)[2],dim(tidy_set)[2],F)
test_i=1:round(0.2*length(index))
train_i=!index \%in\% index[test_i]

## Fit the model
history=
  ontonet \%>\%
  compile(
    loss='mean_squared_error'
    ,loss_weights=c(rep(0.3,length(.$outputs)-1),1)
    ,metrics='accuracy'
  ) \%>\%
  fit_generator(
    generator=
      generator.ontoarray(
        tidy_set
        ,index=index[train_i]
        ,batch_size=4
      )
    ,steps_per_epoch=24
    ,validation_data=
      generator.ontoarray(
        tidy_set
        ,index=index[test_i]
        ,batch_size=4
      )
    ,validation_steps=6
    ,epochs=30
    ,verbose=1
  )
}
\keyword{generator}
\keyword{sample}
