% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_example-function.R
\name{input_example}
\alias{input_example}
\title{Make an input example for divnn package}
\usage{
input_example()
}
\value{
output A list of inputs: 1) value, a data frame with rows for
instances and columns for features; 2) outcome, a vector of binary integers
with the same length as the instances; 3) similarity, a square matrix of
numerics containing feature-feature similarity measures; 4) mapping, a matrix
of numerics with rows for features and three columns for three dimensions
where the features are mapped onto; and 5) ontology, a data frame with rows
for ontologies and four columns for source, target, similarity, and relation.
In addition, a result of hierarchical clustering is also included for
visualization purpose.
}
\description{
This function create an input example for several function in divnn package.
}
\examples{

## Create input example
input=input_example()

## Show output and visualize the ontology by hierarchical clustering
input
plot(input$hierarchy)
}
\keyword{data}
\keyword{example}
