% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontonet_generator.R
\name{ontonet_generator}
\alias{ontonet_generator}
\title{Make an ontonet generator for visible neural network (VNN) modeling}
\usage{
ontonet_generator(tidy_set, path = NULL, init_seed = 888, init2_seed = 9999)
}
\arguments{
\item{tidy_set}{TidySet, an ExpressionSet with three tables.}

\item{path}{A character of file path if the model json file is saved.}

\item{init_seed}{An integer of random seed for ReLU initializer.}

\item{init2_seed}{An integer of random seed for tanh initializer.}
}
\value{
output Keras model object, a pointer to Keras model object in python
environment, which will be an input to train VNN model using Keras R package.
}
\description{
This function create a function that generate a Keras Convolutional Neural
Network (CNN) model with a specific layer architecture for each path in the
hierarchy of the given ontology.
}
\examples{

## Create input example
input=input_example()

## Create a TidySet
tidy_set=
  create_tidy_set(
    value=input$value
    ,outcome=input$outcome
    ,similarity=input$similarity
    ,mapping=input$mapping
    ,ontology=input$ontology
  )

## Create ontonet (Keras model object) generator function
ontonet=ontonet_generator(tidy_set)
}
\keyword{Keras}
\keyword{model}
\keyword{object}
\keyword{ontonet,}
