% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ts_tar_gz-function.R
\name{read_ts_tar_gz}
\alias{read_ts_tar_gz}
\title{Read a .ts.tar.gz file to a TidySet}
\usage{
read_ts_tar_gz(path)
}
\arguments{
\item{path}{A character of .ts.tar.gz file path (include file extension).}
}
\value{
output A TidySet, an ExpressionSet with three tables. Function of
\code{write_ts_tar_gz} can write this file from the TidySet.
}
\description{
This function read multiple files archived by tar with gzip compression
to a TidySet.
}
\examples{

## Create input example
input=input_example()

## Create a TidySet
tidy_set=
  create_tidy_set(
    value=input$value
    ,outcome=input$outcome
    ,similarity=input$similarity
    ,mapping=input$mapping
    ,ontology=input$ontology
  )

## Write a .ts.tar.gz file from a TidySet
write_tidy_set(tidy_set,'example')

## Read a .ts.tar.gz file to a TidySet
read_ts_tar_gz('example.ts.tar.gz')
}
\keyword{.ts.tar.gz,}
\keyword{TidySet}
