% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidySet.compile-function.R
\name{TidySet.compile}
\alias{TidySet.compile}
\title{Make a TidySet for visible neural network (VNN) modeling}
\usage{
TidySet.compile(
  value,
  outcome,
  similarity,
  mapping,
  ontology,
  ranked = T,
  dims = 7,
  decreasing = F,
  seed_num = 33
)
}
\arguments{
\item{value}{Instance-feature value, a data frame with rows for instances
and columns for features. All rows in \code{value} should have names. All
values should be numerics.}

\item{outcome}{Outcome, a vector of binary integers with the same length as
the instances. The length and the order of \code{outcome} should be the same
with those of \code{value}. Value  of 0 and 1 should refer to non-event and
event outcome, respectively.}

\item{similarity}{Feature similarity, a square matrix of numerics containing
feature-feature similarity measures.}

\item{mapping}{Feature three-dimensional mapping, a matrix of integers with
rows for features and three columns for three dimensions where the features
are mapped onto.}

\item{ontology}{Ontology, a data frame with rows for ontologies and four
columns for source, target, similarity, and relation. Feature (source)-
ontology (target) relation should be annotated as 'feature', while ontology-
ontology relation should be annotated as 'is_a'. To differentiate between
feature and ontology names, a prefix of 'ONT:' precedes an ontology name. All
columns except similarity in \code{ontology} should be characters. Similarity
(a numeric) is a minimum threshold by which either features or ontologies
(source) belong to an ontology (target).}
}
\value{
output TidySet, an ExpressionSet with three tables. Instance-feature
value data frame and outcome vector are compiled as a phenotype data frame
with rows for instances and columns for features and outcome. Instance-
feature value data frame and feature three-dimensional mapping matrix are
compiled as an expression matrix with rows for positions of features and
columns for instances. The mapping and similarity matrices and ontology data
frame are compiled as a feature data frame with rows for positions of
features and columns for feature names and ontological relations. For easier,
access the similarity matrix, ontomap four-dimensional array, ontotype list
of two-dimensional matrices, and ontology data frame are included in
experiment notes that can be called using Biobase function \code{notes}.
}
\description{
This function create a TidySet, an ExpressionSet class to orchestrate five
data into a single set of three tables.
}
\examples{

## Create input example
input=utils.example()

## Compile input to a TidySet
tidy_set=
  TidySet.compile(
    value=input$value
    ,outcome=input$outcome
    ,similarity=input$similarity
    ,mapping=input$mapping
    ,ontology=input$ontology
  )

## The TidySet
tidy_set

## The phenotype data frame
pData(tidy_set)

## The feature data frame
fData(tidy_set)

## The expression data frame
exprs(tidy_set)

## Recall a similarity matrix
notes(tidy_set)$similarity

## Recall an ontomap four-dimensional array
notes(tidy_set)$ontomap

## Recall an ontotype list of two-dimensional matrices
notes(tidy_set)$ontotype

## Recall an ontology data frame
notes(tidy_set)$ontology
}
\keyword{ExpressionSet}
\keyword{TidySet,}
