% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidySet.read-function.R
\name{TidySet.read}
\alias{TidySet.read}
\title{Read a .ts.tar.gz file to a TidySet}
\usage{
TidySet.read(path)
}
\arguments{
\item{path}{A character of .ts.tar.gz file path (include file extension).}
}
\value{
output A TidySet, an ExpressionSet with three tables. Function of
\code{TidySet.write} can write this file from the TidySet.
}
\description{
This function read multiple files archived by tar with gzip compression
to a TidySet.
}
\examples{

## Create input example
input=utils.example()

## Compile input to a TidySet
tidy_set=
  TidySet.compile(
    value=input$value
    ,outcome=input$outcome
    ,similarity=input$similarity
    ,mapping=input$mapping
    ,ontology=input$ontology
  )

## Write a .ts.tar.gz file from a TidySet
TidySet.write(tidy_set,'example')

## Read a .ts.tar.gz file to a TidySet
TidySet.read('example.ts.tar.gz')
}
\keyword{.ts.tar.gz,}
\keyword{TidySet}
