% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidySet.write-function.R
\name{TidySet.write}
\alias{TidySet.write}
\title{Write a .ts.tar.gz file from a TidySet}
\usage{
TidySet.write(tidy_set, path)
}
\arguments{
\item{tidy_set}{TidySet, an ExpressionSet with three tables.}

\item{path}{A character of .ts.tar.gz file path (do not include file
extension).}
}
\value{
output A .ts.tar.gz file containing exprs.csv, pData.csv,
fData.csv, similarity.csv, ontology.csv, and others.txt. Function of
\code{TidySet.read} can read this file back to a TidySet.
}
\description{
This function write multiple files archived by tar with gzip compression
from a TidySet.
}
\examples{

## Create input example
input=utils.example()

## Create a TidySet
tidy_set=
  TidySet.create(
    value=input$value
    ,outcome=input$outcome
    ,similarity=input$similarity
    ,mapping=input$mapping
    ,ontology=input$ontology
  )

## Write a .ts.tar.gz file from a TidySet
write_tidy_set(tidy_set,'example')
}
\keyword{.ts.tar.gz,}
\keyword{TidySet}
