% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.ontonet-function.R
\name{viz.ontonet}
\alias{viz.ontonet}
\title{Visualize an ontonet for visible neural network (VNN) modeling}
\usage{
viz.ontonet(
  tidy_set,
  feature = F,
  eval.results = NULL,
  eval.metric = "loss",
  eval.pal = c("darkred", "darkgreen"),
  eval.gradient = 100
)
}
\arguments{
\item{tidy_set}{TidySet, an ExpressionSet with three tables.}

\item{feature}{A logical whether feature nodes are shown; otherwise, only
ontology nodes are shown.}

\item{eval.results}{A list of resampled results or a single result from
\code{evaluate_generator()} of \code{keras} package.}

\item{eval.metric}{A character of metric name to visualize.}

\item{eval.pal}{A vector of two characters for colors representing the
minimum and maximum values of the metric.}

\item{eval.gradient}{An integer of the number of possible colors between
ones for the minimum and maximum values of the metric.}

\item{node.color}{A character of color for bordering node.}

\item{node.fill}{A character of color for filling node.}

\item{node.shape}{A character of node shape. Run \code{vertex.shapes()} from
\code{igraph} package to find the options.}

\item{node.size}{A numeric of node size.}

\item{label}{A logical whether the nodes are labeled.}

\item{label.family}{A character of font family for node label.}

\item{label.cex}{A numeric of font size for node label.}

\item{label.color}{A character of font color for node label.}

\item{edge.color}{A character of color for edge.}

\item{seed_num}{A integer of random seed for network construction algorithm.}
}
\value{
viz.ontonet object, a list of node and edge data frames with the
visualization parameters. Visualization are shown by \code{plot()}.
}
\description{
This function visualizes an ontonet, including the evaluation results for
each ontology.
}
\examples{

## Create input example
input=utils.example()

## Compile input to a TidySet
tidy_set=
  TidySet.compile(
    value=input$value
    ,outcome=input$outcome
    ,similarity=input$similarity
    ,mapping=input$mapping
    ,ontology=input$ontology
  )

## Visualize ontonet
viz.ontonet(tidy_set)
}
\keyword{ontonet}
\keyword{visualization}
